streamlit/requirements.txt
pillow
requests
streamlit
streamlit-lottie
streamlit-option-menu